Comments addressed
