+++ .github/bundles/eks/uds-bundle.yaml
@@ -100,6 +100,8 @@ packages:
               value: false
             - path: autoscaling.enabled
               value: true
+            - path: detailedObservability.alerts.enabled
+              value: true
           variables:
             - name: KEYCLOAK_DB_HOST
               path: postgresql.host
