+++ test/vitest/helpers/alertmanager.ts
@@ -1,6 +1,8 @@
 /**
+ * Copyright 2025 Defense Unicorns
  * SPDX-License-Identifier: AGPL-3.0-or-later OR LicenseRef-Defense-Unicorns-Commercial
  */
+import { pollUntilSuccess } from "./polling";
 
 /**
  * Checks if a specific alert is currently firing in Alertmanager
@@ -31,3 +33,28 @@ export const checkAlertInAlertmanager = async (
     throw new Error(`Failed to query Alertmanager: ${error}`);
   }
 };
+
+/**
+ * Waits for a specific alert to start firing in Alertmanager
+ *
+ * This function polls Alertmanager until the specified alert becomes active or the timeout is reached.
+ * It's useful for integration tests that need to verify alerts are properly configured and firing.
+ *
+ * @param alertmanagerUrl - The base URL of the Alertmanager instance
+ * @param alertName - The name of the alert to wait for
+ * @param timeoutMs - Maximum time to wait for the alert in milliseconds (default: 240000ms / 4 minutes)
+ * @returns Promise that resolves when the alert is firing, or rejects if timeout is reached
+ * @throws Error if the alert doesn't fire within the timeout period
+ */
+export async function expectAlertFires(
+  alertmanagerUrl: string,
+  alertName: string,
+  timeoutMs = 240000, // 4 minutes default timeout (allows time for alert discovery, state change, and safety buffer)
+): Promise<void> {
+  await pollUntilSuccess(
+    () => checkAlertInAlertmanager(alertmanagerUrl, alertName),
+    isAlertFiring => isAlertFiring === true,
+    `Checking for ${alertName} alert in AlertManager`,
+    timeoutMs,
+  );
+}
